requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.19
idna==3.10
pillow==10.4.0
psycopg==3.2.4
requests==2.32.3
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
urllib3==2.2.3
djangorestframework
rest_framework_simplejwt
django-cors-headers
